refactor AppMenu
