+++ packages/react/src/components/AppMenu/Default/Bar.js
@@ -1,0 +1,166 @@
+import React from 'react';
+
+import {
+  AppBar as _AppBar,
+  Burger as _Burger,
+  Button,
+  SocialButtons as _SocialButtons,
+} from '@wingscms/components';
+
+import { t } from '../../../theme';
+import { wrapLink } from '../../../lib/utils';
+import styled from '../../../lib/styled';
+
+const LayoutContainer = styled.div`
+  display: block;
+  position: relative;
+  height: ${t(_ => _.largeSpacing)};
+  width: 100%;
+`;
+
+const BarLayoutContainer = styled.div`
+  position: relative;
+  width: 100%;
+  max-width: 1200px;
+  height: ${t(_ => _.largeSpacing)};
+  padding: 0 ${t(_ => _.smallSpacing)};
+  margin: 0 auto;
+  display: flex;
+  flex-wrap: wrap;
+  flex-direction: row;
+  justify-content: flex-end;
+  align-items: center;
+`;
+
+const BarLayoutRight = styled.div`
+  direction: rtl;
+  height: 100%;
+  margin: 0 0 0 auto;
+  display: flex;
+  flex-wrap: wrap;
+  flex-direction: row;
+  justify-content: flex-start;
+  align-items: center;
+  overflow: hidden;
+  max-width: calc(100% - 220px);
+  white-space: nowrap;
+  @media screen and (max-width: 1300px) {
+    padding-right: ${({ primaryItem }) => (primaryItem ? '0' : '50px')};
+  }
+`;
+
+const BarLayoutRightItem = styled.div`
+  direction: ltr;
+  height: 100%;
+  display: flex;
+  align-items: center;
+`;
+
+const Logo = styled.img`
+  cursor: pointer;
+  max-width: 200px;
+  max-height: calc(100% - 20px);
+  display: block;
+  margin-right: auto;
+  vertical-align: top;
+  justify-self: flex-start;
+`;
+
+const SocialButtons = styled(_SocialButtons)`
+  @media screen and (max-width: 600px) {
+    display: none;
+  }
+`;
+
+const AppBar = styled(_AppBar)`
+  background-color: ${t(_ => _.appMenuBarBackgroundColor)};
+  height: ${t(_ => _.largeSpacing)};
+  box-shadow: ${t(_ => _.shadow)};
+`;
+
+const LangaugeSelectButton = styled(Button)`
+  font-size: 16px;
+  height: 40px;
+  padding: 0 10px;
+  margin-left: 20px;
+  svg {
+    margin: 0;
+  }
+`;
+
+const PrimaryItem = styled(Button)`
+  font-size: 16px;
+  line-height: 40px;
+  padding: 0 20px;
+  margin-left: 10px;
+  @media screen and (max-width: 500px) {
+    display: none;
+  }
+`;
+
+const Burger = styled(_Burger)`
+  position: absolute;
+  right: ${t(_ => _.smallSpacing)};
+  top: 50%;
+  transform: translateY(-50%);
+`;
+
+export default function MenuDefaultBar({
+  burgerProps,
+  logo: { url: logoUrl, alt: logoAlt = 'Logo', wrap: wrapLogo },
+  languageSelectOnClick,
+  menuItems,
+  primaryMenuItems,
+  socialButtons,
+  socialButtonsProps,
+  wrapMenuItem,
+}) {
+  const LogoLink = wrapLink(wrapLogo)(({ children, ...props }) => <a {...props}>{children}</a>);
+  const PrimaryItemLink = wrapLink(wrapMenuItem)(({ children, ...props }) => (
+    <a style={{ textDecoration: 'none' }} {...props}>
+      {children}
+    </a>
+  ));
+
+  return (
+    <AppBar position={AppBar.Position.TOP}>
+      <LayoutContainer>
+        <BarLayoutContainer>
+          {logoUrl ? (
+            <LogoLink href="/">
+              <Logo alt={logoAlt} src={logoUrl} />
+            </LogoLink>
+          ) : null}
+          <BarLayoutRight>
+            <BarLayoutRightItem>
+              <LangaugeSelectButton
+                intent={Button.Intent.PRIMARY}
+                icon="globe"
+                onClick={languageSelectOnClick}
+              />
+            </BarLayoutRightItem>
+            {!primaryMenuItems.length
+              ? null
+              : primaryMenuItems.map(({ url, text, ...props }) => (
+                  <BarLayoutRightItem>
+                    <PrimaryItemLink href={url}>
+                      <PrimaryItem intent={Button.Intent.PRIMARY} {...props}>
+                        {text}
+                      </PrimaryItem>
+                    </PrimaryItemLink>
+                  </BarLayoutRightItem>
+                ))}
+            <BarLayoutRightItem>
+              <SocialButtons>
+                {socialButtons.map(({ platform, url }, i) => (
+                  <SocialButtons.Button {...socialButtonsProps} key={i} icon={platform} url={url} />
+                ))}
+              </SocialButtons>
+            </BarLayoutRightItem>
+          </BarLayoutRight>
+        </BarLayoutContainer>
+        {!menuItems.length ? null : <Burger {...burgerProps} />}
+      </LayoutContainer>
+    </AppBar>
+  );
+}
