+++ packages/react/src/components/AppMenu/Default/Drawer.js
@@ -1,0 +1,91 @@
+import React from 'react';
+
+import {
+  Burger as _Burger,
+  Button,
+  Drawer as _Drawer,
+  Heading,
+  Link,
+  Portal,
+} from '@wingscms/components';
+
+import { t } from '../../../theme';
+import { wrapLink } from '../../../lib/utils';
+import styled from '../../../lib/styled';
+
+const LayoutContainer = styled.div`
+  display: block;
+  position: relative;
+  height: ${t(_ => _.largeSpacing)};
+  width: 100%;
+`;
+
+const Burger = styled(_Burger)`
+  position: absolute;
+  right: ${t(_ => _.smallSpacing)};
+  top: 50%;
+  transform: translateY(-50%);
+`;
+
+const MenuItem = styled.div`
+  ${t(Heading.getStyles(1))}
+  text-align: center;
+  margin-bottom: 1.3em;
+  a {
+    color: ${t(_ => _.appMenuDrawerTextColor)};
+  }
+`;
+
+const Drawer = styled(_Drawer)`
+  background-color: ${t(_ => _.appMenuDrawerBackgroundColor)};
+`;
+
+const PrimaryItemDrawer = styled(Button)`
+  margin: 20px auto;
+  display: block;
+`;
+
+export default function MenuDefaultDrawer({
+  burgerProps,
+  menuItems,
+  primaryMenuItems,
+  wrapMenuItem,
+  open,
+}) {
+  const MenuLink = wrapLink(wrapMenuItem)(Link);
+  const PrimaryItemLink = wrapLink(wrapMenuItem)(({ children, ...props }) => (
+    <a style={{ textDecoration: 'none' }} {...props}>
+      {children}
+    </a>
+  ));
+
+  return !menuItems.length ? null : (
+    <Portal>
+      <Drawer open={open}>
+        <LayoutContainer>
+          <Burger {...burgerProps} />
+        </LayoutContainer>
+        {menuItems.map(item => (
+          <MenuItem>
+            <MenuLink type={Link.Style.LINE_GROW} chref={item.url}>
+              {item.text}
+            </MenuLink>
+          </MenuItem>
+        ))}
+        {!primaryMenuItems.length
+          ? null
+          : primaryMenuItems.map(({ url, text, ...props }) => (
+              <PrimaryItemLink href={url}>
+                <PrimaryItemDrawer
+                  size={Button.Size.SMALL}
+                  intent={Button.Intent.PRIMARY}
+                  {...props}
+                >
+                  {text}
+                </PrimaryItemDrawer>
+              </PrimaryItemLink>
+            ))}
+      </Drawer>
+    </Portal>
+  );
+}
