+++ packages/react/src/lib/utils.js
@@ -1,4 +1,5 @@
 import _slugify from 'slugify';
+import React from 'react';
 
 export const slugify = s =>
   _slugify(s, {
@@ -28,4 +29,19 @@ export const isJSON = item => {
   }
 
   return false;
+};
+
+export const wrapLink = LinkWrap => Link => {
+  if (LinkWrap) {
+    return ({ href, children, ...props }) => (
+      <LinkWrap href={href} {...props}>
+        <Link>{children}</Link>
+      </LinkWrap>
+    );
+  }
+  return ({ href, children, ...props }) => (
+    <Link href={href} {...props}>
+      {children}
+    </Link>
+  );
 };
